Create chrono.gs
This is chrono, as found https://github.com/wanasit/chrono . Google scripts didn't allow to reference by URC or use the reduced version of the source, so the exploded version was coppied and pasted to this file. Updates to chrono should be found here https://github.com/wanasit/chrono

All credit and thanks to wanasit for it.